gpt_disk_io: Move slice_block_io into the block_io module
